Fix to Mike's warning fixes
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@33833 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/mesh/animesh/controllers/ragdoll/ragdoll.cpp
@@ -461,7 +461,7 @@ CS_PLUGIN_NAMESPACE_BEGIN(Ragdoll)
       childNode->BlendState (state, baseWeight);
 
     // reset the chains that have been asked for
-    for (size_t i = resetChains.GetSize () - 1; i >= 0; i--)
+    for (int i = ((int) resetChains.GetSize ()) - 1; i >= 0; i--)
     {
       ResetChainData &chainData = resetChains.Get (i);
       chainData.frameCount--;
